mcp-router/requirements.txt
fastmcp==2.13.1
mcp==1.21.0
jsonschema==4.25.1
pytest-asyncio==0.21.0
aiohttp==3.13.2
PyYAML==6.0.3
